refactor(b64): move the b64 impl into the jose-b64 crate
