+++ jose-b64/Cargo.toml
@@ -14,6 +14,13 @@ readme = "README.md"
 edition = "2021"
 rust-version = "1.63"
 
+[dependencies]
+serde = { version = "1.0.136", default-features = false, features = ["alloc", "derive"] }
+zeroize = { version = "1.5.7", default-features = false, features = ["alloc", "serde"] }
+subtle = { version = "2.4.1", default-features = false }
+
+serde_json = { version = "1.0.79", default-features = false, optional = true, features = ["alloc"] }
+
 [package.metadata.docs.rs]
 all-features = true
 rustdoc-args = ["--cfg", "docsrs"]
