+++ jose-b64/src/json.rs
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2022 Profian Inc. <opensource@profian.com>
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
-#![cfg(feature = "jws")]
+#![cfg(feature = "serde_json")]
 
 use alloc::boxed::Box;
 use alloc::vec::Vec;
