+++ jose-b64/src/optional.rs
