+++ jose/Cargo.toml
@@ -5,7 +5,7 @@ edition = "2021"
 
 [features]
 jwk = []
-jws = ["p256?/ecdsa", "p384?/ecdsa", "dep:serde_json"]
+jws = ["p256?/ecdsa", "p384?/ecdsa", "dep:serde_json", "jose-b64/serde_json"]
 sha2 = ["dep:sha2", "dep:digest"]
 hmac = ["dep:rand_core", "dep:hmac"]
 p256 = ["dep:rand_core", "dep:p256"]
@@ -16,6 +16,7 @@ rsa = ["dep:rand_core", "dep:rsa"]
 serde = { version = "1.0.136", default-features = false, features = ["alloc", "derive"] }
 zeroize = { version = "1.5.7", default-features = false, features = ["alloc", "serde"] }
 subtle = { version = "2.4.1", default-features = false }
+jose-b64 = { path = "../jose-b64" }
 jose-jwa = { path = "../jose-jwa" }
 
 # Optional Dependencies
